[TRL-234] refactor: 사용자 정의 리포지토리 정의 및 기존 Repository 의존 관계 변경
