+++ src/main/java/com/cosain/trilo/config/JpaConfig.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @Configuration
 @EnableJpaRepositories(
-        value = {"com.cosain.trilo.trip.command.domain", "com.cosain.trilo.user.domain"}
+        value = {"com.cosain.trilo.trip.command.domain", "com.cosain.trilo.user.domain", "com.cosain.trilo.trip.query.infra.repository.trip.jpa"}
 )
 public class JpaConfig {
 }
