+++ src/main/java/com/cosain/trilo/trip/query/infra/repository/trip/TripQueryRepositoryImpl.java
@@ -1,7 +1,8 @@
 package com.cosain.trilo.trip.query.infra.repository.trip;
 
-import com.cosain.trilo.trip.query.infra.dto.TripDetail;
 import com.cosain.trilo.trip.query.domain.repository.TripQueryRepository;
+import com.cosain.trilo.trip.query.infra.dto.TripDetail;
+import com.cosain.trilo.trip.query.infra.repository.trip.jpa.TripQueryJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -11,10 +12,10 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class TripQueryRepositoryImpl implements TripQueryRepository {
 
-    private final TripQueryDslRepository tripQueryDslRepository;
+    private final TripQueryJpaRepository tripQueryJpaRepository;
 
     @Override
     public Optional<TripDetail> findTripDetailByTripId(Long tripId) {
-        return tripQueryDslRepository.findTripDetailById(tripId);
+        return tripQueryJpaRepository.findTripDetailById(tripId);
     }
 }
