+++ src/test/java/com/cosain/trilo/unit/trip/query/infra/repository/TripQueryRepositoryTest.java
@@ -1,31 +1,26 @@
 package com.cosain.trilo.unit.trip.query.infra.repository;
 
-import com.cosain.trilo.config.QueryDslConfig;
+import com.cosain.trilo.support.RepositoryTest;
 import com.cosain.trilo.trip.command.domain.entity.Trip;
 import com.cosain.trilo.trip.command.domain.vo.TripPeriod;
 import com.cosain.trilo.trip.command.domain.vo.TripStatus;
+import com.cosain.trilo.trip.query.domain.repository.TripQueryRepository;
 import com.cosain.trilo.trip.query.infra.dto.TripDetail;
-import com.cosain.trilo.trip.query.infra.repository.trip.TripQueryDslRepository;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.context.annotation.Import;
 
 import java.time.LocalDate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@Import({QueryDslConfig.class, TripQueryDslRepository.class})
-@DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@DisplayName("TripQueryDslRepository 테스트")
-public class TripQueryRepositoryImplTest {
+@RepositoryTest
+@DisplayName("TripQueryRepository 테스트")
+public class TripQueryRepositoryTest {
 
     @Autowired
-    private TripQueryDslRepository tripQueryDslRepository;
+    private TripQueryRepository tripQueryRepository;
 
     @Autowired
     private EntityManager em;
@@ -42,7 +37,7 @@ public class TripQueryRepositoryImplTest {
         em.persist(trip);
 
         // when
-        TripDetail tripDetail = tripQueryDslRepository.findTripDetailById(1L).get();
+        TripDetail tripDetail = tripQueryRepository.findTripDetailByTripId(1L).get();
 
         // then
         assertThat(tripDetail.getTitle()).isEqualTo(trip.getTitle());
